requirements.txt
pynput>=1.7.6
gspread>=5.7.0
google-auth>=2.20.0
